Masony and Lazy Load Rework: - Rework lazy load script to optionally use masonry for laying out images. - Rework masonry initialization for gallery preview and main gallery. - Update js imports.
