+++ apps/common/templates/base.html
@@ -72,7 +72,7 @@
             <script src="{% static 'js/main.js' %}"></script>
             <script src="{% static 'js/ToTopButton.js' %}"></script>
             <script src="{% static 'js/ThemeSwitcher/ThemeSwitcher.js' %}"></script>
-            <script src="{% static 'js/LazyLoader/LazyLoader.js' %}"></script>
+            <script type="module" src="{% static 'js/galleryMasonry/galleryMasonry.js' %}"></script>
             <script src="https://unpkg.com/imagesloaded@5/imagesloaded.pkgd.min.js"></script>
             <script src="https://unpkg.com/masonry-layout@4/dist/masonry.pkgd.min.js"></script>
         {% endblock %}
