+++ apps/gallery/templates/gallery_detail.html
@@ -62,13 +62,18 @@
                     title="{{ photo.title }} on {{ photo.date_taken }}"
                     href="{{ photo.get_display_url }}"
                     data-lightbox="{{ gallery }}"
-                    class="d-block"
                   >
-                    <img
-                      class="travel-photo d-block"
-                      src="{{ photo.get_display_url }}"
-                      alt="{{ photo.title }}: {{ photo.caption }}"
-                    >
+                    <span class="lazy-image-wrapper" data-masonry-item>
+                      <div class="spinner"></div>
+                      <img
+                        class="lazy-image travel-photo d-block"
+                        src="{{ photo.get_thumbnail_url }}"
+                        data-src="{{ photo.get_display_url }}"
+                        data-lazy="image"
+                        data-duration="2000"
+                        alt="{{ photo.title }}: {{ photo.caption }}"
+                      >
+                    </span>
                   </a>
                 </div>
               {% endfor %}
@@ -84,5 +89,5 @@
 {% block extra_script_links %}
   <script src="https://cdn.jsdelivr.net/npm/bootstrap-select@1.13.14/dist/js/bootstrap-select.min.js"></script>
   <script src="{% static 'lightbox/js/lightbox-plus-jquery.js' %}"></script>
-  <script src="{% static 'js/gallery_detail.js' %}"></script>
+  <script type="module" src="{% static 'js/gallery_detail.js' %}"></script>
 {% endblock %}
