+++ apps/index/templates/partials/gallery.html
@@ -31,5 +31,6 @@
   </div>
 </section>
 
+{% block extra_script_links %}
 <script src="{% static 'lightbox/js/lightbox-plus-jquery.js' %}"></script>
-<script src="https://unpkg.com/masonry-layout@4/dist/masonry.pkgd.min.js"></script>
+{% endblock %}
